+++ core/src/main/java/com/predic8/membrane/core/interceptor/apikey/stores/ApiKeyStore.java
@@ -1,8 +1,7 @@
-package com.predic8.membrane.core.interceptor.apikey.store;
+package com.predic8.membrane.core.interceptor.apikey.stores;
 
 import java.util.List;
 
 public interface ApiKeyStore {
-
     List<String> getScopes(String apiKey);
 }
